refactored to have just one manual and ci script
